fix: refactor inmemory to develop
