+++ src/main/java/ru/yandex/practicum/filmorate/storage/friendlist/FriendListDbStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.storage.friendlist;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.model.FriendsList;
+import ru.yandex.practicum.filmorate.storage.BaseStorage;
 
 import java.util.List;
