+++ src/main/java/ru/yandex/practicum/filmorate/storage/friendlist/FriendListStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.storage.friendlist;
 
 import ru.yandex.practicum.filmorate.model.FriendsList;
